+++ src/app/AppModule.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { CoreModule } from '@/core/CoreModule';
 import { TodoModule } from '@/todo/TodoModule';
 import { AppConfigModule } from '@/config/AppConfigModule';
+import { DatabaseModule } from '@/database/DatabaseModule';
 
 @Module({
-  imports: [CoreModule, TodoModule, AppConfigModule],
+  imports: [AppConfigModule, DatabaseModule, CoreModule, TodoModule],
   controllers: [],
   providers: [],
 })
