+++ src/config/AppConfigModule.ts
@@ -1,18 +1,13 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { DataBaseConfigService } from '@/config/DataBaseConfigService';
 import { ConfigModule } from '@nestjs/config';
-import database from '@/config/database';
+import { DatabaseConfig } from '@/config/database';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       // must be global for any module to use load typeorm repository correctly
       isGlobal: true,
-      load: [database],
-    }),
-    TypeOrmModule.forRootAsync({
-      useClass: DataBaseConfigService,
+      load: [DatabaseConfig],
     }),
   ],
   controllers: [],
