+++ src/database/DatabaseModule.ts
@@ -1,0 +1,12 @@
+import { DataBaseConfigService } from '@/config/database';
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+@Module({
+  imports: [
+    TypeOrmModule.forRootAsync({
+      useClass: DataBaseConfigService,
+    }),
+  ],
+})
+export class DatabaseModule {}
